+++ src/mimic/api/src/schema.rs
@@ -1,6 +1,6 @@
 use crate::Error;
 use candid::CandidType;
-use orm_schema::node::Canister;
+use orm_schema::node::{Canister, CanisterBuild};
 use serde::{Deserialize, Serialize};
 use snafu::Snafu;
 
@@ -20,6 +20,18 @@ pub enum SchemaError {
 ///
 /// SCHEMA FUNCTIONS
 ///
+
+// canisters_by_build
+pub fn canisters_by_build(build: CanisterBuild) -> Result<Vec<Canister>, Error> {
+    let schema = core_schema::get_schema().map_err(SchemaError::from)?;
+    let canisters: Vec<Canister> = schema
+        .filter_nodes::<Canister, _>(|canister| canister.build == build)
+        .map(|(_, v)| v)
+        .cloned()
+        .collect();
+
+    Ok(canisters)
+}
 
 // canister
 pub fn canister(path: &str) -> Result<Canister, Error> {